add_bot_send_group_logs; add_html_to telegram
